+++ background.js
@@ -4,6 +4,8 @@
 const PORT_SETTINGS = "settings";
 const MODE_ALL=1, MODE_TAB=2, MODE_URL=3;
 const CSP_ENABLE=1, CSP_RESTRICT=2, CSP_DISABLE=3;
+const THEME_LIGHT=1, THEME_DARK=2;
+const REG_DOMAINURL = /^[\.\*\?a-z0-9_-]+$/;
 const BROWSERACTION_TITLE = {true: "ScriptFilter (OFF)", false: "ScriptFilter (ON)"};
 const BROWSERACTION_ICON = {true: "/images/icon-cspoff.png", false: "/images/icon-cspon.png"};
 const CONTEXTMENU_OPTION = "option", CONTEXTMENU_PAGE = "page", CONTEXTMENU_LINK = "link";
@@ -12,7 +14,6 @@ const CONTEXTMENU_PAGE_TITLE = {
 	false: geti18ndata({key: "contextmenu_page2", msg: "Disable CSP"})
 };
 const CONTEXTMENU_LINK_TITLE = geti18ndata({key: "contextmenu_link", msg: "Open link in new tab with CSP"});
-const REG_DOMAINURL = /^[\.\*\?a-z0-9_-]+$/;
 const CSP_SANDBOX_NOPOPUPS = "sandbox allow-forms allow-orientation-lock allow-pointer-lock"
 	+" allow-presentation allow-same-origin allow-scripts allow-top-navigation;";
 
@@ -23,9 +24,9 @@ function CSPManager(){
 	this.csp = {};
 	this.mode = MODE_TAB;
 	this.options = {};
+	this.ui = {};
 	this.urls = [];
 	this.trust = true;
-	this.currentid = null;
 	this.settingsid = null;
 	var self = this;
 	var untrusts = [];
@@ -47,8 +48,12 @@ function CSPManager(){
 		self.options = storage.options || {};
 		self.options = {
 			contextmenu: (!ANDROID && self.options.contextmenu!=false),
-			urlsadd: (self.options.urlsadd!=false),
-			urlsdel: (self.options.urlsdel!=false)
+			autoaddurl: (self.options.autoaddurl!=false),
+			autodelurl: (self.options.autodelurl!=false)
+		};
+		self.ui = storage.ui || {
+			theme: THEME_LIGHT,
+			settings: {section: null, exporttype: null}
 		};
 		self.urls = storage.urls || [];
 		self.trust = (self.mode!=MODE_ALL || storage.trust!=false);
@@ -60,11 +65,6 @@ function CSPManager(){
 	//Gets current url from focused tab
 	browser.tabs.query({currentWindow: true, active: true},function(tabs){
 		if(tabs.length) self.setcurrent(tabs.first());
-	});
-	
-	//Tab close
-	browser.tabs.onRemoved.addListener(function(tabid,info){
-		untrusts.remove(val => (val==tabid));
 	});
 	
 //#	Public methods
@@ -92,7 +92,7 @@ function CSPManager(){
 	this.setmode = function(val){
 		this.mode = parseInt(val);
 		if(this.mode==MODE_URL) untrusts = untrusts.concat(this.urls.slice(0));
-		else untrusts.removeAll(val => !Number.isInteger(val));
+		else untrusts = untrusts.filter(Number.isInteger);
 		if(this.options.contextmenu) updatecontextmenu(CONTEXTMENU_LINK);
 		this.trust = true;
 		this.updatebutton();
@@ -106,6 +106,12 @@ function CSPManager(){
 		browser.storage.local.set({options: this.options});
 	};
 	
+	this.setui = function(page,obj){
+		if(obj.uitheme) this.ui.theme = obj.uitheme;
+		else for(var prop in obj.uistate) this.ui[page][prop] = obj.uistate[prop];
+		browser.storage.local.set({ui: this.ui});
+	};
+	
 	//Adds url to blacklist
 	this.addurl = function(url){
 		if(url.match(REG_DOMAINURL) && !this.urls.contains(url)){
@@ -114,47 +120,50 @@ function CSPManager(){
 			browser.storage.local.set({urls: this.urls});}
 	};
 	
+	//Adds all urls to blacklist
+	this.addurlall = function(list){
+		var urls = list.unique().filter(url => (url && url.match(REG_DOMAINURL) && !this.urls.find(val => urlmatch(val,url))));
+		this.urls = this.urls.concat(urls);
+		if(this.mode==MODE_URL) untrusts = untrusts.concat(urls);
+		browser.storage.local.set({urls: this.urls});
+		return urls.length;
+	};
+	
 	//Removes url from blacklist
 	this.delurl = function(url){
 		this.urls.remove(val => (val==url));
-		if(this.mode==MODE_URL) untrusts.removeAll(val => (val==url));
-		browser.storage.local.set({urls: this.urls});
-	};
-	
-	//Imports blacklist
-	this.importurls = function(urls){
-		urls.removeAll(val => (!val.match(REG_DOMAINURL) || this.urls.contains(val)));
-		this.urls = this.urls.concat(urls);
-		if(this.mode==MODE_URL) untrusts = untrusts.concat(urls);
+		if(this.mode==MODE_URL) untrusts = untrusts.filter(val => (val!=url));
 		browser.storage.local.set({urls: this.urls});
 	};
 	
 	//Clears blacklist
 	this.clear = function(){
+		var length = this.urls.length;
 		this.urls = [];
-		if(this.mode==MODE_URL) untrusts.removeAll(val => !Number.isInteger(val));
+		if(this.mode==MODE_URL) untrusts = untrusts.filter(Number.isInteger);
 		browser.storage.local.set({urls: this.urls});
+		return length;
 	};
 	
 	//Reloads all untrusted tabs
 	this.reload = function(){
-		untrusts
-			.filter(val => Number.isInteger(val))
-			.forEach(val => browser.tabs.reload(val,{bypassCache: true}));
+		browser.tabs.query({},function(tabs){
+			untrusts = untrusts.filter(val => (!Number.isInteger(val) || tabs.find(tab => (tab.id==val))));
+			tabs.filter(tab => (tab.id!=self.settingsid))
+				.forEach(tab => browser.tabs.reload(tab.id,{bypassCache: true}));
+		});
 	};
 	
 	//Returns true if tab is trusted
 	this.istrusted = function(tab){
-		if(this.settingsid==tab.id) return true;
 		switch(this.mode){
 		case MODE_ALL : return this.trust;
-		case MODE_TAB : return !untrusts.find(val => (val==tab.id));
-		case MODE_URL : return !untrusts.find(val => urlmatch(val,urltodomain(tab.url)));}
+		case MODE_TAB : return (this.settingsid==tab.id || !untrusts.find(val => (val==tab.id)));
+		case MODE_URL : return (this.settingsid==tab.id || !untrusts.find(val => urlmatch(val,urltodomain(tab.url))));}
 	};
 	
-	//Sets current focused tab
+	//Sets current focused or updated tab
 	this.setcurrent = function(tab){
-		this.currentid = tab.id;
 		switch(this.mode){
 		case MODE_TAB :
 			this.trust = !untrusts.find(val => (val==tab.id));
@@ -162,9 +171,9 @@ function CSPManager(){
 		break;
 		case MODE_URL :
 			var domainurl = urltodomain(tab.url);
-			if(!domainurl) break;
-			this.trust = !untrusts.find(val => urlmatch(val,domainurl));
+			this.trust = !domainurl || !untrusts.find(val => urlmatch(val,domainurl));
 			this.updatebutton();
+			if(!domainurl) break;
 		case MODE_ALL :
 			var cspoff = !untrusts.find(val => (val==tab.id));
 			if(!this.trust && cspoff && this.settingsid!=tab.id){
@@ -178,7 +187,7 @@ function CSPManager(){
 	
 	//Sets trust to browser/tab/domain
 	this.settrust = function(tab){
-		if(this.settingsid!=tab.id || this.mode==MODE_ALL){
+		if(this.mode==MODE_ALL || this.settingsid!=tab.id){
 			this.trust = true;
 			switch(this.mode){
 			case MODE_ALL : browser.storage.local.set({trust: this.trust});
@@ -187,16 +196,15 @@ function CSPManager(){
 			break;
 			case MODE_URL :
 				var domainurl = urltodomain(tab.url);
-				if(this.options.urlsdel) this.urls.remove(val => (val==domainurl));
-				untrusts.removeAll(val => (val==tab.id || urlmatch(val,domainurl)));
+				if(this.options.autodelurl) this.urls.remove(val => (val==domainurl));
+				untrusts = untrusts.filter(val => (val!=tab.id && !urlmatch(val,domainurl)));
 				browser.storage.local.set({urls: this.urls});
-			break;}
-			this.updatebutton();}
+			break;}}
 	};
 	
 	//Sets untrust to browser/tab/domain
 	this.setuntrust = function(tab){
-		if(this.settingsid!=tab.id || this.mode==MODE_ALL){
+		if(this.mode==MODE_ALL || this.settingsid!=tab.id){
 			this.trust = false;
 			switch(this.mode){
 			case MODE_ALL : browser.storage.local.set({trust: this.trust});
@@ -205,20 +213,20 @@ function CSPManager(){
 			break;
 			case MODE_URL :
 				var domainurl = urltodomain(tab.url);
-				if(!domainurl) break;
-				var val = this.urls.find(val => urlmatch(val,domainurl));
-				if(!val && this.options.urlsadd) this.urls.push(domainurl);
-				untrusts.push(tab.id);
-				untrusts.push(val || domainurl);
-				browser.storage.local.set({urls: this.urls});
-			break;}
-			this.updatebutton();}
+				if(domainurl){
+					var val = this.urls.find(val => urlmatch(val,domainurl));
+					if(!val && this.options.autoaddurl) this.urls.push(domainurl);
+					untrusts.push(tab.id);
+					untrusts.push(val || domainurl);
+					browser.storage.local.set({urls: this.urls});}
+			break;}}
 	};
 	
 	//Switches Content Security Policy ON/OFF
 	this.toggletrust = function(tab){
 		if(this.trust) this.setuntrust(tab);
 		else this.settrust(tab);
+		this.updatebutton();
 		if(this.settingsid!=tab.id) browser.tabs.reload(tab.id,{bypassCache: true});
 	};
 	
@@ -229,27 +237,30 @@ function CSPManager(){
 		if(!ANDROID) browser.browserAction.setIcon({path: BROWSERACTION_ICON[this.trust]});
 		if(this.options.contextmenu) updatecontextmenu(CONTEXTMENU_PAGE);
 	};
-	
-//#	Private methods
-	var csptostr = function(key,val){
-		switch(val){
-		case CSP_ENABLE : return "";
-		case CSP_RESTRICT : return key+" 'self';";
-		case CSP_DISABLE : return key+" 'none';";}
-	};
-	
-	var urltodomain = function(url){
-		return (url.match(/:\/\/(.[^/]+)/) || [])[1];
-	};
-	
-	var urlmatch = function(val1,val2){
-		return (/[\*\?]/.test(val1) && val2)?
-			val2.match(val1.replaceAll(".","\\.").replaceAll("*",".*").replaceAll("?",".")):
-			(val1==val2);
-	};
 }
 
 /* -------------------- Functions -------------------- */
+
+//Converts CSP key+value to string
+function csptostr(key,val){
+	switch(val){
+	case CSP_ENABLE : return "";
+	case CSP_RESTRICT : return key+" 'self';";
+	case CSP_DISABLE : return key+" 'none';";}
+}
+
+//Converts URL to domain name
+function urltodomain(url){
+	var result = url.toLowerCase().match(/https?:\/\/(.[^\/]+)/);
+	return (result && result[1] || "");
+}
+
+//Returns true if the 2 domain urls matches
+function urlmatch(url1,url2){
+	return (/[\*\?]/.test(url1) && url2)?
+		url2.match("^"+url1.replaceAll(".","\\.").replaceAll("*",".*").replaceAll("?",".")+"$"):
+		(url1==url2);
+}
 
 //Creates context menu items
 function createcontextmenu(){
@@ -264,28 +275,97 @@ function createcontextmenu(){
 		contexts: ["link"],
 		enabled: (cspman.mode!=MODE_ALL)
 	});
-};
+}
 
 //Updates context menu item
 function updatecontextmenu(menuid){
 	switch(menuid){
-	case CONTEXTMENU_PAGE :
-		browser.contextMenus.update(CONTEXTMENU_PAGE,{title: CONTEXTMENU_PAGE_TITLE[cspman.trust]});
+	case CONTEXTMENU_PAGE : browser.contextMenus.update(CONTEXTMENU_PAGE,{title: CONTEXTMENU_PAGE_TITLE[cspman.trust]});
 	break;
-	case CONTEXTMENU_LINK :
-		browser.contextMenus.update(CONTEXTMENU_LINK,{enabled: (cspman.mode!=MODE_ALL)});
+	case CONTEXTMENU_LINK : browser.contextMenus.update(CONTEXTMENU_LINK,{enabled: (cspman.mode!=MODE_ALL)});
 	break;}
-};
+}
 
 //Removes context menu items
 function removecontextmenu(){
 	browser.contextMenus.remove(CONTEXTMENU_PAGE);
 	browser.contextMenus.remove(CONTEXTMENU_LINK);
-};
+}
 
 //Returns true if HTTP request is opened by Chromium PDF Viewer
 function ischromiumpdf(info){
 	return (CHROMIUM && info.responseHeaders.find(val => (val.name=="Content-Type" && val.value=="application/pdf")));
+}
+
+//Adds all tab urls to blacklist
+function blacklistall(){
+	browser.tabs.query({},function(tabs){
+		var urls = tabs.map(tab => urltodomain(tab.url));
+		portcon.send({port: PORT_SETTINGS, msg: {status: "blacklistall", result: cspman.addurlall(urls)}});
+		portcon.send(PORT_SETTINGS);
+	});
+}
+
+//Sets trust to all open tabs
+function trustall(){
+	if(cspman.mode!=MODE_ALL)
+		browser.tabs.query({},function(tabs){
+			tabs.filter(tab => !cspman.istrusted(tab))
+				.forEach(function(tab){
+					cspman.settrust(tab);
+					browser.tabs.reload(tab.id,{bypassCache: true});
+				});
+			portcon.send(PORT_SETTINGS);
+		});
+	else{
+		cspman.settrust();
+		cspman.updatebutton()};
+}
+
+//Sets untrust to all open tabs
+function untrustall(){
+	if(cspman.mode!=MODE_ALL)
+		browser.tabs.query({},function(tabs){
+			tabs.filter(tab => (cspman.settingsid!=tab.id && cspman.istrusted(tab)))
+				.forEach(function(tab){
+					cspman.setuntrust(tab);
+					browser.tabs.reload(tab.id,{bypassCache: true});
+				});
+			portcon.send(PORT_SETTINGS);
+		});
+	else{
+		cspman.setuntrust();
+		cspman.updatebutton();}
+}
+
+//Gets settings data received from port message
+function getsettingsdata(msg){
+	switch(msg.status){
+	case "csp" : cspman.setcsp(msg.csp);
+	break;
+	case "mode" : cspman.setmode(msg.mode);
+	break;
+	case "options" : cspman.setoptions(msg.options);
+	break;
+	case "ui" : cspman.setui(PORT_SETTINGS,msg);
+	break;
+	case "addurl" : cspman.addurl(msg.url);
+	break;
+	case "delurl" : cspman.delurl(msg.url);
+	break;
+	case "import" : portcon.send({port: PORT_SETTINGS, msg: {status: "import", result: cspman.addurlall(msg.urls)}});
+	break;
+	case "clear" : portcon.send({port: PORT_SETTINGS, msg: {status: "clear", result: cspman.clear()}});
+	break;
+	case "blacklistall" : blacklistall();
+	break;
+	case "untrustall" : untrustall();
+	break;
+	case "trustall" : trustall();
+	break;
+	case "reload" : cspman.reload();
+	break;}
+	portcon.send(PORT_SETTINGS);
 }
 
 /* -------------------- Main Process -------------------- */
@@ -297,7 +377,7 @@ var cspman = new CSPManager();
 //Browser action button click
 browser.browserAction.onClicked.addListener(function(tab){
 	cspman.toggletrust(tab);
-	portcon.send([PORT_SETTINGS]);
+	portcon.send(PORT_SETTINGS);
 });
 
 //OptionsPage browserAction context menu
@@ -321,13 +401,15 @@ browser.contextMenus.onClicked.addListener(function(info,tab){
 	break;
 	case CONTEXTMENU_PAGE :
 		cspman.toggletrust(tab);
-		portcon.send([PORT_SETTINGS]);
+		portcon.send(PORT_SETTINGS);
 	break;
 	case CONTEXTMENU_LINK :
-		opentab({url: info.linkUrl, index: "next"},function(tab){
+		opentab({url: info.linkUrl, index: "next", active: false},function(tab){
 			tab.url = info.linkUrl;
-			if(cspman.istrusted(tab)) cspman.setuntrust(tab);
-			portcon.send([PORT_SETTINGS]);
+			if(cspman.istrusted(tab)){
+				cspman.setuntrust(tab);
+				cspman.updatebutton();}
+			portcon.send(PORT_SETTINGS);
 		});
 	break;}
 });
@@ -347,7 +429,7 @@ browser.tabs.onActivated.addListener(function(info){
 
 //Tab update
 browser.tabs.onUpdated.addListener(function(tabid,info,tab){
-	if(cspman.currentid==tabid) cspman.setcurrent(tab);
+	cspman.setcurrent(tab);
 });
 
 //Blocking web requests with Content Security Policy
@@ -370,32 +452,18 @@ browser.runtime.onConnect.addListener(function(port){
 	portcon.connect({
 		port: port,
 		msgpost: function(){ return {
+			status: "settings",
 			csp: cspman.csp,
 			mode: cspman.mode,
 			options: cspman.options,
 			urls: cspman.urls.sort()
 		}},
-		msgget: function(msg){
-			switch(msg.status){
-			case "CSP" : cspman.setcsp(msg.csp);
-			break;
-			case "mode" : cspman.setmode(msg.mode);
-			break;
-			case "options" : cspman.setoptions(msg.options);
-			break;
-			case "addurl" : cspman.addurl(msg.url);
-			break;
-			case "delurl" : cspman.delurl(msg.url);
-			break
-			case "import" : cspman.importurls(msg.urls);
-			break
-			case "clear" : cspman.clear();
-			break
-			case "reload" : cspman.reload();
-			break;}
-			portcon.send([PORT_SETTINGS]);
-		},
+		msgget: getsettingsdata,
 		disconnect: function(){ cspman.settingsid = null }
 	});
-	portcon.send([PORT_SETTINGS]);
+	portcon.send({
+		port: PORT_SETTINGS,
+		msg: {status: "ui", uitheme: cspman.ui.theme, uistate: cspman.ui[PORT_SETTINGS]}
+	});
+	portcon.send(PORT_SETTINGS);
 });